refactor: cleanup signatures
